padamo-package/requirements.txt
numpy
numba
numba-scipy @ git+https://github.com/numba/numba-scipy.git@1e2f244
scipy @ git+https://github.com/scipy/scipy.git@7f2ac69
h5py
h5pickle
matplotlib
pillow
tqdm
toposort
imageio
imageio[ffmpeg]
pyinstaller
paramiko
appdirs
serde
psutil

# Cuda is shipped with tensorflow! Finally!
tensorflow[and-cuda]
uproot
tbb
